backend/requirements.txt
# Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Real-time
websockets==12.0
python-socketio==5.10.0

# Integrations
aiohttp==3.9.1
httpx==0.25.2
openai==1.3.8
boto3==1.29.7
google-auth==2.26.1
google-auth-oauthlib==1.2.0

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78

# Database Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production
gunicorn==21.2.0

# Utilities
python-multipart==0.0.6
